Versión final aplicativo 26-05-19

+++ AssistMeProject/AssistMeProject/Controllers/UsersController.cs
@@ -1,15 +1,17 @@
-﻿using System;
+﻿using AssistMeProject.Models;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Http;
+using System.Collections;
+using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using AssistMeProject.Models;
-using System.Security.Claims;
-using System.Threading;
-using System.Web;
+using System.Net.Mail;
+using System.Text;
 
 namespace AssistMeProject.Controllers
 {
@@ -28,14 +30,7 @@ namespace AssistMeProject.Controllers
         // GET: Users
         public IActionResult Index(string message)
         {
-            setActiveUser();
-            List<SelectListItem> list = new List<SelectListItem>();
-            var studios = _context.Studio.ToList();
-            foreach (Studio s in studios)
-            {
-                list.Add(new SelectListItem() { Text = s.Name, Value = s.Name });
-            }
-            ViewBag.Studios = new SelectList(list, "Value", "Text");
+            string userS = SetActiveUser();
             ViewBag.MESSAGE = message;
             return View();
         }
@@ -43,7 +38,9 @@ namespace AssistMeProject.Controllers
         // GET: Users/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            setActiveUser();
+            string userS = SetActiveUser();
+            if (string.IsNullOrEmpty(userS))
+                return RedirectToAction("", "", new { message = "Error, Inicie sesión" });
             if (id == null)
             {
                 return NotFound();
@@ -102,7 +99,9 @@ namespace AssistMeProject.Controllers
         // GET: Users/Edit/5
         public IActionResult Edit()
         {
-            setActiveUser();
+            string userS = SetActiveUser();
+            if (string.IsNullOrEmpty(userS))
+                return RedirectToAction("Index", "Users", new { message = "Error, Inicie sesión" });
             string currentlyActiveUsername = HttpContext.Session.GetString(ACTIVE_USERNAME);
 
             if (currentlyActiveUsername == null)
@@ -127,8 +126,10 @@ namespace AssistMeProject.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit([Bind("ID,GOOGLE_KEY,LEVEL,USERNAME,PASSWORD,EMAIL,PHOTO,QUESTIONS_ANSWERED,POSITIVE_VOTES_RECEIVED,QUESTIONS_ASKED,INTERESTING_VOTES_RECEIVED,DESCRIPTION,INTERESTS_OR_KNOWLEDGE,COUNTRY,CITY,StudioId")] User user)
         {
-            setActiveUser();
-            if (ModelState.IsValid && _context.User.Count(p => p.USERNAME.Equals(user.USERNAME)) == 1)
+            
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(ACTIVE_USERNAME)))
+                return RedirectToAction("Index", "Users", new { message = "Error, Inicie sesión" });
+            if (ModelState.IsValid && CheckUsername(user.USERNAME))
             {
                 try
                 {
@@ -152,10 +153,17 @@ namespace AssistMeProject.Controllers
             return View(user);
         }
 
+        private bool CheckUsername(string username)
+        {
+            return _context.User.Count(p => p.USERNAME.Equals(username)) == 1;
+        }
+
         // GET: Users/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            setActiveUser();
+            string userS = SetActiveUser();
+            if (string.IsNullOrEmpty(userS))
+                return RedirectToAction("", "", new { message = "Error, Inicie sesión" });
             if (id == null)
             {
                 return NotFound();
@@ -176,18 +184,13 @@ namespace AssistMeProject.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            setActiveUser();
+            string userS = SetActiveUser();
+            if (string.IsNullOrEmpty(userS))
+                return RedirectToAction("", "", new { message = "Error, Inicie sesión" });
             var user = await _context.User.FindAsync(id);
             _context.User.Remove(user);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private void setActiveUser()
-        {
-            //To pass the username active
-            ViewBag.ACTIVE_USER = HttpContext.Session.GetString(ACTIVE_USERNAME);
-            //End To pass the username active
         }
 
         private bool UserExists(int id)
@@ -203,7 +206,9 @@ namespace AssistMeProject.Controllers
         [HttpGet]
         public IActionResult Profile(string viewingToUser)
         {
-            setActiveUser();
+            string userS = SetActiveUser();
+            if (string.IsNullOrEmpty(userS))
+                return RedirectToAction("Index", "Users", new { message = "Error, Inicie sesión" });
             string currentlyActiveUsername = HttpContext.Session.GetString(ACTIVE_USERNAME);
 
             if (!string.IsNullOrEmpty(currentlyActiveUsername))
@@ -211,7 +216,10 @@ namespace AssistMeProject.Controllers
                 if (string.IsNullOrEmpty(viewingToUser))
                     return View(model.GetUser(currentlyActiveUsername));
                 ViewData["ACTIVE_USER"] = currentlyActiveUsername;
-                return View(model.GetUser(viewingToUser));
+                User found = model.GetUser(viewingToUser);
+                found.QUESTIONS_ASKED = _context.Question.Count(a => a.UserId == found.ID);//We profile added this because it wasn't implemented action of increase or decrease when a user makes or answer a question
+                found.QUESTIONS_ANSWERED = _context.Answer.Count(a => a.UserId == found.ID);
+                return View(found);
             }
             else
             {
@@ -222,8 +230,9 @@ namespace AssistMeProject.Controllers
         [HttpPost]
         public IActionResult Profile(string username, string password, string method)
         {
-            setActiveUser();
             User found = model.FindUser(username, password, method);
+            found.QUESTIONS_ASKED = _context.Question.Count(a => a.UserId==found.ID); //We profile added this because it wasn't implemented action of increase or decrease when a user makes or answer a question
+            found.QUESTIONS_ANSWERED = _context.Answer.Count(a => a.UserId == found.ID);
             if (found == null)
             {
                 return RedirectToAction("Index", "Users", new { message = "Error, prueba de nuevo" });
@@ -233,6 +242,7 @@ namespace AssistMeProject.Controllers
                 //Only username it's saved for have a better security but this might be slower because have to search user every time it's needed
                 HttpContext.Session.SetString(ACTIVE_USERNAME, found.USERNAME);
                 ViewData["ACTIVE_USER"] = username;
+                SetActiveUser();
                 return View(found);
             }
         }
@@ -241,6 +251,71 @@ namespace AssistMeProject.Controllers
         {
             HttpContext.Session.Remove(ACTIVE_USERNAME);
             return RedirectToAction("Index", "Users");
+        }
+
+        public IActionResult AllNotifications()
+        {
+            string userActive = SetActiveUser();
+            if (string.IsNullOrEmpty(userActive))
+            {
+                return RedirectToAction("Index", "Users", new { message = "Error, Inicie sesión" });
+            }
+            else
+            {
+                int id = model.GetUser(userActive).ID;
+                var notifications = _context.Notification.Where(p => p.UserID == id).ToList();
+                GetNotificationsOfUser();
+                return View(notifications);
+            }
+
+
+        }
+
+        private void GetNotificationsOfUser()
+        {
+            string userActive = HttpContext.Session.GetString(ACTIVE_USERNAME);
+            User user = model.GetUser(userActive);
+            try
+            {
+                ViewBag.Notifications = _context.Notification.Where(p => p.UserID == user.ID && !p.Read).ToList();
+            } catch (Exception e)
+            {
+
+            }
+        }
+
+        public IActionResult LostPassword(string email)
+        {
+            User found = _context.User.FirstOrDefault(a => a.EMAIL.Equals(email));
+            if (found==null)
+                return RedirectToAction("Index", "Users", new { message = "Errror, no se encontro ningún usuario registrado con el correo: " + email });
+            string message = "<br><h2>Tu nombre de usuario es: <span style=\"color:#f40000\"><bold>" + found.USERNAME+ "</bold></span><br>Y tu contraseña es: <span style=\"color:#f40000\"><bold>" + found.PASSWORD+ "</bold></span></h2>" +
+                "<br><br>Gracias por formar parte de la familia Globant";
+            try
+            {
+                Email manager = new Email();
+                manager.EnviarCorreo(email, "Recuperar contraseña", message,true);
+            } catch(Exception e)
+            {
+                Console.WriteLine(e.Message);
+            }
+            return RedirectToAction("Index","Users",new { message = "Se envió su nombre de usuario y contraseña"});
+        }
+
+        /**
+         * This method allow to set the name of the active user. If there is no user, then pass the Studios that exist for create an account
+         **/
+        private String SetActiveUser()
+        {
+            //To pass the username active
+            string USER = HttpContext.Session.GetString(ACTIVE_USERNAME);
+            if (string.IsNullOrEmpty(USER))
+                ViewBag.Studios = AssistMe.GetSelectListStudios(_context);
+            else
+                GetNotificationsOfUser();
+            ViewBag.ACTIVE_USER = USER;
+            return USER;
+            //End To pass the username active
         }
 
     }
